+++ qom/ui/log.py
@@ -45,7 +45,7 @@ def init_log(log_format='full', debug=False):
     main_logger.addHandler(handler)
 
     # test
-    logger.info('-------------------Logger Initialized-------------------\n')
+    logger.info('-------------------------------Logger Initialized-------------------\n')
     
 def get_formatter(log_format='full'):
     """Function to obtain the formatter for stream handler.